requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
pytz>=2024.1
python-dotenv>=1.0.0
requests>=2.31.0
